Clean up renderer header // Static cast file util
[git-p4: depot-paths = "//depot/C35/Students/cwsullivan/SD/": change = 449203]

+++ Engine/Code/Engine/Core/FileUtils.cpp
@@ -37,5 +37,5 @@ int FileReadToString( std::string& outString, const std::string& filename )
 	FileReadToBuffer(outBuffer, filename);
 	outBuffer.push_back('\0');
 	outString = std::string( outBuffer.begin(), outBuffer.end() );
-	return outString.size();
+	return static_cast<int>(outString.size());
 }
